Stop updating npm for rhel

+++ lib/vagrant-openshift/action/install_origin_asset_dependencies.rb
@@ -35,8 +35,11 @@ ASSET_BACKUP_DIR=/data/asset_dependencies
 
 if ! which npm > /dev/null 2>&1 ; then
   sudo yum -y install npm
-  # Update npm, we need a newer version than is installed by yum
-  sudo npm install npm@3.7.3 -g
+  if $(test -e /etc/fedora-release)
+  then
+    # Update npm, we need a newer version than is installed by yum
+    sudo npm install npm@3.7.3 -g
+  fi
 fi
 
 }
